refactor: OtherUserPostをOtherUserSakuraCardに変更
