+++ Front/front-react-project/src/OtherUserSakuraCard.tsx
